+++ src/payment/models/staking/staking.service.ts
@@ -4,14 +4,14 @@ import { Deposit } from '../deposit/deposit.entity';
 import { DepositService } from '../deposit/deposit.service';
 import { Sell } from '../sell/sell.entity';
 import { SellRepository } from '../sell/sell.repository';
-import { KycCompleted } from '../../../user/models/userData/userData.entity';
+import { KycCompleted } from '../../../user/models/user-data/user-data.entity';
 import { CreateStakingDto } from './dto/create-staking.dto';
 import { StakingType } from './dto/staking-type.enum';
 import { StakingDto } from './dto/staking.dto';
 import { UpdateStakingDto } from './dto/update-staking.dto';
 import { Staking } from './staking.entity';
 import { StakingRepository } from './staking.repository';
-import { UserDataService } from 'src/user/models/userData/userData.service';
+import { UserDataService } from 'src/user/models/user-data/user-data.service';
 import { CryptoInputRepository } from '../crypto-input/crypto-input.repository';
 import { AssetService } from 'src/shared/models/asset/asset.service';
 import { Asset } from 'src/shared/models/asset/asset.entity';
