+++ src/user/models/ident/ident.service.ts
@@ -1,12 +1,12 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CountryService } from 'src/shared/models/country/country.service';
 import { Util } from 'src/shared/util';
-import { KycInProgress, KycState, KycStatus, UserData } from 'src/user/models/userData/userData.entity';
+import { KycInProgress, KycState, KycStatus, UserData } from 'src/user/models/user-data/user-data.entity';
 import { KycDocument } from 'src/user/services/kyc/dto/kyc.dto';
 import { KycService, KycProgress } from 'src/user/services/kyc/kyc.service';
-import { AccountType } from '../userData/userData.entity';
-import { UserDataRepository } from '../userData/userData.repository';
-import { UserDataService } from '../userData/userData.service';
+import { AccountType } from '../user-data/user-data.entity';
+import { UserDataRepository } from '../user-data/user-data.repository';
+import { UserDataService } from '../user-data/user-data.service';
 import { IdentUserDataDto } from './dto/ident-user-data.dto';
 
 export interface KycResult {
