+++ src/user/models/user-data/user-data.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { UserData } from './userData.entity';
+import { UserData } from './user-data.entity';
 
 @EntityRepository(UserData)
 export class UserDataRepository extends Repository<UserData> {}
