+++ src/user/models/user/dto/user.dto.ts
@@ -1,7 +1,7 @@
 import { Fiat } from 'src/shared/models/fiat/fiat.entity';
 import { Language } from 'src/shared/models/language/language.entity';
-import { AccountType } from '../../userData/userData.entity';
-import { KycState, KycStatus } from '../../userData/userData.entity';
+import { AccountType } from '../../user-data/user-data.entity';
+import { KycState, KycStatus } from '../../user-data/user-data.entity';
 import { UserStatus } from '../user.entity';
 
 export interface UserDto {
