+++ src/user/models/user/user.service.spec.ts
@@ -3,7 +3,7 @@ import { createMock } from '@golevelup/ts-jest';
 import { CountryService } from 'src/shared/models/country/country.service';
 import { FiatService } from 'src/shared/models/fiat/fiat.service';
 import { LanguageService } from 'src/shared/models/language/language.service';
-import { UserDataService } from '../userData/userData.service';
+import { UserDataService } from '../user-data/user-data.service';
 import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
 import { KycService } from 'src/user/services/kyc/kyc.service';
