+++ src/user/user.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AinModule } from 'src/ain/ain.module';
 import { SharedModule } from 'src/shared/shared.module';
-import { UserDataController } from 'src/user/models/userData/userData.controller';
-import { UserDataRepository } from 'src/user/models/userData/userData.repository';
-import { UserDataService } from 'src/user/models/userData/userData.service';
+import { UserDataController } from 'src/user/models/user-data/user-data.controller';
+import { UserDataRepository } from 'src/user/models/user-data/user-data.repository';
+import { UserDataService } from 'src/user/models/user-data/user-data.service';
 import { WalletRepository } from 'src/user/models/wallet/wallet.repository';
 import { WalletService } from 'src/user/models/wallet/wallet.service';
 import { AuthController } from './models/auth/auth.controller';
